Fixed bug where timestep data contained the same Numpy arrays for the states and the next states, causing the model to never learn how to move towards helpful states.
Changed the timesteps back down to a reasonable number, and batch size
back up over 1.

Some things to think about:
1) is it possible to remove state information from the Agent class so
that only the GridWorld contains it? That way not two things are
referring to the general idea of game state.

2) When I was probing the issue, I checked to see the epsilon values and
noticed that when the model wasn't learning, epsilon dropped too low too
quickly, like down to the floor after 2 games. That seems too fast but
maybe for this problem which is simple enough, we don't need it to
explore because there's not much more information to be gained through
exploration?