Added grid width and height as arguments. Not fully tested, but everything seems to work correctly when those arguments are not specified. I think the value mapping function needs to be tweaked to accommodate different dimensions.
Changed reward mapping function to do "sampling" instead of averaging
equally between all actions that could go to a state.

Made reward mapping function calculate reward for the same grid that it
was trained on. I don't know if this is necessary but it would reduce
the complexity of the issues we're dealing with right now.